% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dCRT}
\alias{dCRT}
\title{The distilled conditional randomization test.}
\usage{
dCRT(
  X,
  Y,
  Z,
  family,
  method = list(XZ = "glm", YZ = "glm"),
  fitted.ext = list(XZ = NULL, YZ = NULL),
  alternative = "two.sided",
  B = 5000
)
}
\arguments{
\item{X}{Numeric vector of length \eqn{n}, representing the predictor variable.}

\item{Y}{Numeric vector of length \eqn{n}, representing the response variable.}

\item{Z}{Numeric matrix with \eqn{n} rows and \eqn{p} columns, representing covariates.}

\item{family}{Named list with elements \code{XZ} and \code{YZ} specifying the model family for \eqn{X \mid Z} and
\eqn{Y \mid Z} for each model. Each list element must be a string (e.g. \code{"binomial"},
\code{"poisson"}). Ignored for any model where you supply your own fitted values via \code{fitted.ext}.}

\item{method}{Named list with elements \code{XZ} and \code{YZ} that selects the model-fitting method to use
for each model. Each element must be a string (e.g. \code{"glm"}, \code{"rf"}, \code{"prob_forest"}).
Ignored for any model where you supply your own fitted values via \code{fitted.ext}.}

\item{fitted.ext}{Named list with elements \code{XZ} and \code{YZ} of user-supplied fitted values
(numeric vectors of length n). For each non-NULL element, that model is
treated as custom and neither \code{method} nor \code{family} is used.}

\item{alternative}{A character string specifying the alternative hypothesis for the test.
Values can be "two.sided" (default), "greater", or "less".}

\item{B}{The number of resamples to draw (Default value is 2000).}
}
\value{
An named list containing the following fields:
\describe{
\item{test_stat}{The test statistic.}
\item{p_value}{The p-value under the specified alternative.}
}
}
\description{
\code{dCRT} is a function carrying out the dCRT based on GLMs for \code{X|Z} and \code{Y|Z}.
}
\examples{
n <- 200; p <- 4
set.seed(1234)
X <- rbinom(n = n, size = 1, prob = 0.3)
Y <- rpois(n = n, lambda = 1)
Z <- matrix(rnorm(n = n*p, mean = 0, sd = 1), nrow = n, ncol = p)

# fit both models via GLM
res.dCRT.1 <- dCRT(X = X, Y = Y, Z = Z,
                   family = list(XZ = "binomial", YZ = "poisson"),
                   method = list(XZ = "glm", YZ = "glm"))
res.dCRT.1

# custom fit for Y|Z, explicit method/family only for X|Z
user_fit_Y <- glm(Y ~ Z,
                  family = poisson(),
                  data = data.frame(Y = Y, Z = Z))$fitted.values |> unname()

res.dCRT.2 <- dCRT(X = X, Y = Y, Z = Z,
                   method = list(XZ = "rf"),
                   family = list(XZ = "binomial"),
                   fitted.ext = list(XZ = NULL, YZ = user_fit_Y),
                   alternative = "greater")
res.dCRT.2

}
\seealso{
\code{\link{model_families}}
}
