% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{fit_models}
\alias{fit_models}
\title{\code{fit_models} is a function carrying out the regressions \code{X} on \code{Z} and \code{Y} on \code{Z}.}
\usage{
fit_models(
  data,
  X_on_Z_fam,
  Y_on_Z_fam,
  fitting_X_on_Z = "glm",
  fitting_Y_on_Z = "glm",
  fit_vals_X_on_Z_own = NULL,
  fit_vals_Y_on_Z_own = NULL
)
}
\arguments{
\item{data}{A (non-empty) named list with fields \code{X} (an nx1 vector for the predictor
variable of interest), \code{Y} (an nx1 response vector), and \code{Z} (an nxp matrix of covariates).}

\item{X_on_Z_fam}{The GLM family for the regression of X on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{Y_on_Z_fam}{The GLM family for the regression of Y on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{fitting_X_on_Z}{The fitting method for the regression X on Z
(values can be \code{glm} (default), \code{rf}, \code{prob_forest}, or \code{own}).}

\item{fitting_Y_on_Z}{The fitting method for the regression Y on Z
(values can be \code{glm} (default), \code{rf}, \code{prob_forest}, or \code{own}).}

\item{fit_vals_X_on_Z_own}{Vector of fitted values for X on Z in case the user's custom method.
Works only if fitting_X_on_Z = 'own'.}

\item{fit_vals_Y_on_Z_own}{Vector of fitted values for Y on Z in case the user's custom method.
Works only if fitting_Y_on_Z = 'own'.}
}
\value{
A named list of fitted values of X|Z and Y|Z.
}
\description{
\code{fit_models} is a function carrying out the regressions \code{X} on \code{Z} and \code{Y} on \code{Z}.
}
\keyword{internal}
