% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{fit_single_model}
\alias{fit_single_model}
\title{\code{fit_single_model} is a function carrying out a single regression \code{V} on \code{Z}, where
\code{V} is the variable/response and \code{Z} is the covariate.}
\usage{
fit_single_model(V, Z, V_on_Z_fam, fitting_V_on_Z, fit_vals_V_on_Z_own)
}
\arguments{
\item{V}{Vector of variable/response.}

\item{Z}{Matrix of covariates.}

\item{V_on_Z_fam}{The GLM family for the regression of V on Z
(values can be \code{gaussian}, \code{binomial}, \code{poisson}, \code{negative.binomial}, etc).}

\item{fitting_V_on_Z}{The fitting method for the regression V on Z
(values can be \code{glm} (default) or \code{random_forest}).}

\item{fit_vals_V_on_Z_own}{Vector of fitted values for V on Z in case the user's custom method.
Works only if fitting_V_on_Z = 'own'.}
}
\value{
A vector of fitted values of V|Z.
}
\description{
\code{fit_single_model} is a function carrying out a single regression \code{V} on \code{Z}, where
\code{V} is the variable/response and \code{Z} is the covariate.
}
\keyword{internal}
