% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_models.R
\name{supported_models}
\alias{supported_models}
\title{Supported Model Families and Fitting Methods for Inference}
\description{
Supported Families and Model-fitting Methods
}
\details{
This page documents the families and fitting methods supported in the \code{X | Z} and
\code{Y | Z} models for functions \code{\link{GCM}}, \code{\link{dCRT}}, and \code{\link{spaCRT}}.

\strong{Supported Families for \code{X | Z}:}
\itemize{
\item \code{"binomial"}
\item \code{"poisson"}
}

Other families may be added in future versions.

\strong{Supported Families for \code{Y | Z}:}
\itemize{
\item \code{"binomial"}
\item \code{"poisson"}
\item \code{"negative.binomial"}
}

Other families may be added in future versions.

\strong{Supported Fitting Methods:}
\itemize{
\item \code{"glm"}
\item \code{"random_forest"}
}

Other methods may be added in future versions.
}
\seealso{
\code{\link{GCM}}, \code{\link{dCRT}}, \code{\link{spaCRT}}
}
\keyword{documentation}
